refactor: 💡 Move map_tx_builder_error_2_str to util/other.rs
